requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
flask-openapi3==2.5.2
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
pytz==2023.3
SQLAlchemy-Utils==0.41.1
